agent/pyproject.toml
[project]
name = "wedding-planner-agent"
version = "0.1.0"
description = "AI wedding planning agent powered by LangGraph, LangChain, and LangSmith"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langgraph>=0.3.0",
    "langsmith>=0.2.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
    "supabase>=2.0.0",
]

[tool.setuptools]
packages = ["wedding_agent"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]
